+++ DocMAH.UnitTests/BaseTestFixture.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using DocMAH.Configuration;
 using Moq;
 using NUnit.Framework;
 
@@ -13,7 +14,10 @@ namespace DocMAH.UnitTests
 	{
 		#region Protected Properties
 
+		protected MockConfiguration Configuration { get; set; }
+
 		protected MockRepository Mocks { get; set; }
+
 		protected ModelFactory Models { get; set; }		
 
 		#endregion
@@ -24,12 +28,14 @@ namespace DocMAH.UnitTests
 		public void BaseSetUp()
 		{
 			Mocks = new MockRepository(MockBehavior.Strict);
+			Configuration = new MockConfiguration(Mocks);
 			Models = new ModelFactory();
 		}
 
 		[TearDown]
 		public void BaseTearDown()
 		{
+			Configuration = null;
 			Mocks = null;
 			Models = null;
 		}