+++ DocMAH/Configuration/IDocmahConfiguration.cs
@@ -7,7 +7,15 @@ namespace DocMAH.Configuration
 	public interface IDocmahConfiguration
 	{
 		string ConnectionStringName { get; set; }
-		string JsUrl { get; set; }
+
 		string CssUrl { get; set; }
+
+		IDocumentationConfiguration DocumentationConfiguration { get; set; }
+
+		IEditHelpConfiguration EditHelpConfiguration { get; set; }
+
+		string JsUrl { get; set; }
+
+		IPopupViewerConfiguration PopupViewerConfiguration { get; set; }
 	}
 }