+++ DocMAH/Configuration/IHidePopupButtonConfiguration.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace DocMAH.Configuration
+{
+	public interface IHidePopupButtonConfiguration
+	{
+		string Description { get; set; }
+
+		bool IsHidden { get; set; }
+
+		string Text { get; set; }
+	}
+}